requirements.txt
django>=5.2
python-dotenv
django-jazzmin
requests
djangorestframework
drf-yasg
django-cors-headers
python-dotenv
